backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
google-generativeai==0.7.2
duckduckgo-search==4.4.0
aiofiles==23.2.1
python-multipart==0.0.9
openai==1.12.0
openai-agents==0.8.0
aiohttp
